backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2
PyJWT
python-dotenv
pytz
sqlparse
tzdata
whitenoise
django-recaptcha